更新关注status取值判断/私聊websocket路径

+++ frontend/h5/publish.js
@@ -3,10 +3,20 @@ const path = require('path');
 const srcDir = path.resolve(__dirname, 'dist');
 const srcCommonsDir = path.resolve(__dirname, './src/js/commons');
 const srcStylesDir = path.resolve(__dirname, './src/style');
-const destDir = path.resolve(__dirname, '../../weibar/web/src/main/resources/static/static');
+
+/**
+ *目录路径 
+  weibar_fontend\frontend\h5
+  weibar\weibar\web\src\main\resources
+ */
+const destDir = path.resolve(__dirname, '../../../weibar/weibar/web/src/main/resources/static/static');
+if(!fsE.existsSync(destDir)){
+  console.log('目录路径不正确。');
+  return false;
+}
 fsE.emptyDirSync(path.join(destDir, 'js'));
 fsE.copySync( srcCommonsDir, path.join(destDir, 'js', 'commons'));
 fsE.copySync( srcStylesDir, path.join(destDir, 'style'));
 fsE.copy( srcDir, destDir)
   .then(() => console.log('success!'))
-  .catch(err => console.error(err))
+  .catch(err => console.error(err));
